Created feature_dog
