--- domain/src/main/java/uk/co/jamiecruwys/domain/model/DogImageDomainModel.kt
@@ -1,5 +1,0 @@
-package uk.co.jamiecruwys.domain.model
-
-data class DogImageDomainModel(
-    val imageUrl: String,
-)
