--- domain/src/main/java/uk/co/jamiecruwys/domain/repository/DogImageRepository.kt
@@ -1,7 +1,0 @@
-package uk.co.jamiecruwys.domain.repository
-
-import uk.co.jamiecruwys.domain.model.DogImageDomainModel
-
-interface DogImageRepository {
-    suspend fun getRandomImage(): DogImageDomainModel?
-}
