--- domain/src/main/java/uk/co/jamiecruwys/domain/retrofit/DogImageResponse.kt
@@ -1,15 +1,0 @@
-package uk.co.jamiecruwys.domain.retrofit
-
-import uk.co.jamiecruwys.domain.model.DogImageDomainModel
-
-typealias DogImageResponse = List<String>
-
-internal fun DogImageResponse.toDomainModel(): DogImageDomainModel {
-    // Data source comes back with .mp4 and .webm which isn't supported by Glide.
-    // Let's filter the data to only have entries that we can display
-    val filteredItems =
-        this.filter {
-            it.endsWith(".png") || it.endsWith(".jpg") || it.endsWith(".jpeg") || it.endsWith(".gif")
-        }
-    return DogImageDomainModel(imageUrl = "https://random.dog/" + filteredItems.random())
-}
