--- domain/src/main/java/uk/co/jamiecruwys/domain/retrofit/DogImageRetrofitService.kt
@@ -1,8 +1,0 @@
-package uk.co.jamiecruwys.domain.retrofit
-
-import retrofit2.http.GET
-
-interface DogImageRetrofitService {
-    @GET("doggos")
-    suspend fun getRandomImageAsync(): DogImageResponse?
-}
