--- domain/src/main/java/uk/co/jamiecruwys/domain/usecase/GetDogImageUseCase.kt
@@ -1,27 +1,0 @@
-package uk.co.jamiecruwys.domain.usecase
-
-import uk.co.jamiecruwys.domain.repository.DogImageRepository
-import java.io.IOException
-import javax.inject.Inject
-
-class GetDogImageUseCase
-    @Inject
-    constructor(
-        private val dogImageRepository: DogImageRepository,
-    ) {
-        sealed class Result {
-            data class Success(val imageUrl: String) : Result()
-
-            data class Error(val e: Throwable) : Result()
-        }
-
-        suspend fun execute(): Result {
-            return try {
-                dogImageRepository.getRandomImage()?.let {
-                    Result.Success(it.imageUrl)
-                } ?: Result.Error(RuntimeException("No image url"))
-            } catch (e: IOException) {
-                Result.Error(e)
-            }
-        }
-    }
