--- domain/src/test/java/uk/co/jamiecruwys/domain/DomainModuleExampleUnitTest.kt
@@ -1,11 +1,0 @@
-package uk.co.jamiecruwys.domain
-
-import org.junit.jupiter.api.Assertions.assertEquals
-import org.junit.jupiter.api.Test
-
-class DomainModuleExampleUnitTest {
-    @Test
-    fun addition_isCorrect() {
-        assertEquals(4, 2 + 2)
-    }
-}
