+++ feature_dog/src/main/AndroidManifest.xml
