+++ feature_dog/src/main/java/uk/co/jamiecruwys/dog/data/DogImageRepository.kt
@@ -1,0 +1,7 @@
+package uk.co.jamiecruwys.dog.data
+
+import uk.co.jamiecruwys.dog.domain.model.DogImage
+
+interface DogImageRepository {
+    suspend fun getRandomImage(): DogImage?
+}
