+++ feature_dog/src/main/java/uk/co/jamiecruwys/dog/data/datasource/api/service/DogImageRetrofitService.kt
@@ -1,0 +1,9 @@
+package uk.co.jamiecruwys.dog.data.datasource.api.service
+
+import retrofit2.http.GET
+import uk.co.jamiecruwys.dog.data.datasource.api.response.DogImageResponse
+
+interface DogImageRetrofitService {
+    @GET("doggos")
+    suspend fun getRandomImageAsync(): DogImageResponse?
+}
