+++ feature_dog/src/main/java/uk/co/jamiecruwys/dog/domain/model/DogImage.kt
@@ -1,0 +1,5 @@
+package uk.co.jamiecruwys.dog.domain.model
+
+data class DogImage(
+    val imageUrl: String,
+)
