+++ navigation/src/main/java/uk/co/jamiecruwys/navigation/CreateNavigationRoutes.kt
@@ -3,7 +3,7 @@ package uk.co.jamiecruwys.navigation
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.compose.composable
 import uk.co.jamiecruwys.showcase.ui.screen.catsScreen
-import uk.co.jamiecruwys.showcase.ui.screen.dogsScreen
+import uk.co.jamiecruwys.dog.presentation.screen.dogsScreen
 import uk.co.jamiecruwys.showcase.ui.screen.homeScreen
 
 fun NavGraphBuilder.createNavigationRoutes(uiEventNavigator: UiEventNavigator) {
